Move expectAreaCurve to its own file
