+++ test/helpers/expectAreaCurve.ts
@@ -1,0 +1,14 @@
+import { expect } from 'vitest';
+import { assertNotNull } from '../helper/assertNotNull';
+
+type ExpectedArea = {
+  d: string;
+};
+
+export function expectAreaCurve(container: Element, expectedAreas: ReadonlyArray<ExpectedArea>) {
+  assertNotNull(container);
+  const areaCurves = container.querySelectorAll('.recharts-area-curve');
+  assertNotNull(areaCurves);
+  const actualAreas = Array.from(areaCurves).map(area => ({ d: area.getAttribute('d') }));
+  expect(actualAreas).toEqual(expectedAreas);
+}
